Migrate `GameDataRes::sub_50DDAE`

+++ src/table.h
@@ -417,7 +417,7 @@ public:
     int sub_5126A2(uint32_t type_low_byte, uint32_t face); // Get monster table index by type and face
     Item* sub_510EE6(uint16_t item_id); // Create a new item instance by item data ID
     int sub_50DF19(int experience); // Get random recruitable monster index for inn reward
-    int sub_50DDAE(uint32_t type_id, uint32_t face); // Get monster data index from encoded item type/face
+    int sub_50DDAE(uint8_t type_id, uint8_t face); // Get monster data index from encoded item type/face
 
     uint8_t sub_50D8BA(CString* name, CString* out); // 50D8BA: Find shape index by item name string
     uint8_t sub_50DA04(CString* name, CString* out); // 50DA04: Find material index by item name string
